+++ benches/ironoxide_bench.rs
@@ -1,6 +1,6 @@
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use futures::executor::block_on;
-use ironoxide::{document::advanced::*, prelude::*};
+use ironoxide::prelude::*;
 use tokio::runtime::Runtime;
 
 /// Setup for dev1 environment
